move Rendering modules to subdirectory like formula
