+++ arelle/ViewFileRenderedLayout.py
@@ -12,8 +12,8 @@ from arelle.FunctionXs import xsString
 from arelle.ModelObject import ModelObject
 from arelle.ModelFormulaObject import Aspect, aspectModels, aspectRuleAspects, aspectModelAspect, aspectStr
 from arelle.PythonUtil import OrderedSet
-from arelle.RenderingResolution import resolveTableStructure
-from arelle.RenderingLayout import layoutTable
+from arelle.rendering.RenderingResolution import resolveTableStructure
+from arelle.rendering.RenderingLayout import layoutTable
 from arelle import XbrlConst
 from arelle.XmlUtil import elementFragmentIdentifier, addQnameValue
