+++ arelle/rendering/RenderingEvaluator.py
