+++ arelle/rendering/RenderingResolution.py
